requirements.txt
dbt-core
dbt-snowflake
pandas
requests
python-dotenv
snowflake-connector-python
